backend/requirements.txt
fastapi
uvicorn[standard]
networkx
pyjwt
bcrypt
python-dotenv
sqlalchemy
email-validator
